# workload/tpcc: make initial data generation totally deterministic