--- a/pkg/workload/tpcc/generate.go
+++ b/pkg/workload/tpcc/generate.go
@@ -78,6 +78,7 @@ var itemColTypes = []types.T{
 func (w *tpcc) tpccItemInitialRowBatch(rowIdx int, cb coldata.Batch, a *bufalloc.ByteAllocator) {
 	l := w.localsPool.Get().(*generateLocals)
 	defer w.localsPool.Put(l)
+	l.rng.Seed(w.seed + uint64(rowIdx))
 
 	iID := rowIdx + 1
 
@@ -119,6 +120,7 @@ func (w *tpcc) tpccWarehouseInitialRowBatch(
 ) {
 	l := w.localsPool.Get().(*generateLocals)
 	defer w.localsPool.Put(l)
+	l.rng.Seed(w.seed + uint64(rowIdx))
 
 	wID := rowIdx // warehouse ids are 0-indexed. every other table is 1-indexed
 
@@ -174,6 +176,7 @@ var stockColTypes = []types.T{
 func (w *tpcc) tpccStockInitialRowBatch(rowIdx int, cb coldata.Batch, a *bufalloc.ByteAllocator) {
 	l := w.localsPool.Get().(*generateLocals)
 	defer w.localsPool.Put(l)
+	l.rng.Seed(w.seed + uint64(rowIdx))
 
 	sID := (rowIdx % numStockPerWarehouse) + 1
 	wID := (rowIdx / numStockPerWarehouse)
@@ -244,6 +247,7 @@ func (w *tpcc) tpccDistrictInitialRowBatch(
 ) {
 	l := w.localsPool.Get().(*generateLocals)
 	defer w.localsPool.Put(l)
+	l.rng.Seed(w.seed + uint64(rowIdx))
 
 	dID := (rowIdx % numDistrictsPerWarehouse) + 1
 	wID := (rowIdx / numDistrictsPerWarehouse)
@@ -313,6 +317,7 @@ func (w *tpcc) tpccCustomerInitialRowBatch(
 ) {
 	l := w.localsPool.Get().(*generateLocals)
 	defer w.localsPool.Put(l)
+	l.rng.Seed(w.seed + uint64(rowIdx))
 
 	cID := (rowIdx % numCustomersPerDistrict) + 1
 	dID := ((rowIdx / numCustomersPerDistrict) % numDistrictsPerWarehouse) + 1
@@ -331,7 +336,7 @@ func (w *tpcc) tpccCustomerInitialRowBatch(
 	if cID <= 1000 {
 		lastName = randCLastSyllables(cID-1, a)
 	} else {
-		lastName = randCLast(l.rng, a)
+		lastName = w.randCLast(l.rng, a)
 	}
 
 	cb.Reset(customerColTypes, 1)
@@ -407,6 +412,7 @@ var historyColTypes = []types.T{
 func (w *tpcc) tpccHistoryInitialRowBatch(rowIdx int, cb coldata.Batch, a *bufalloc.ByteAllocator) {
 	l := w.localsPool.Get().(*generateLocals)
 	defer w.localsPool.Put(l)
+	l.rng.Seed(w.seed + uint64(rowIdx))
 
 	// This used to be a V4 uuid made through the normal `uuid.MakeV4`
 	// constructor, but we 1) want them to be deterministic and 2) want these rows