--- a/pkg/workload/tpcc/order_status.go
+++ b/pkg/workload/tpcc/order_status.go
@@ -136,10 +136,10 @@ func (o *orderStatus) run(ctx context.Context, wID int) (interface{}, error) {
 	// 2.6.1.2: The customer is randomly selected 60% of the time by last name
 	// and 40% by number.
 	if rng.Intn(100) < 60 {
-		d.cLast = string(randCLast(rng, &o.a))
+		d.cLast = string(o.config.randCLast(rng, &o.a))
 		atomic.AddUint64(&o.config.auditor.orderStatusByLastName, 1)
 	} else {
-		d.cID = randCustomerID(rng)
+		d.cID = o.config.randCustomerID(rng)
 	}
 
 	tx, err := o.mcp.Get().BeginEx(ctx, o.config.txOpts)