Ideally, each workload's initial data set would be totally deterministic
from the flags. We didn't do this originally for tpcc because random
number seeding was so expensive that it was showing up as a significant
cpu user in profiles. The golang.org/x/exp/rand package has much faster
seeding though, so get this property back.

Tiny slowdown in benchmarks, but at this point, initial data is nowhere
near the bottleneck for fixtures import. I'm totally okay with a 1-2%
hit for deterministic data.

    name                             old time/op    new time/op    delta
    InitialData/tpcc/warehouses=1-8     357ms ± 2%     361ms ± 2%  +1.19%  (p=0.000 n=25+25)

    name                             old speed      new speed      delta
    InitialData/tpcc/warehouses=1-8   309MB/s ± 2%   305MB/s ± 2%  -1.16%  (p=0.000 n=25+25)

    name                             old alloc/op   new alloc/op   delta
    InitialData/tpcc/warehouses=1-8     193kB ± 0%     193kB ± 0%    ~     (all equal)

    name                             old allocs/op  new allocs/op  delta
    InitialData/tpcc/warehouses=1-8       592 ± 0%       592 ± 0%    ~     (all equal)

Release note: None