@danhhz 2019-06-18
> Looks like teamcity is getting a different fingerprint than I am locally (though different teamcity builds seem to get getting the same value as each other)

This was the result of a couple calls to the global rand that I accidentally made in address.go

bors r=tbg